Refactor database management: migrate from cache to persistent MySQL storage
- Updated import paths for databaseManager in socketController, messageController, and eventHandler to reflect new location.
- Introduced a new databaseManager module that handles message, event, group, contact, and chat storage in MySQL.
- Implemented automatic data retention and cleanup features.
- Added methods for saving and retrieving messages, events, groups, contacts, and chats.
- Enhanced error handling and logging throughout the database operations.
