+++ app/controllers/messageController.js
@@ -10,7 +10,7 @@
  */
 
 require('dotenv').config();
-const { databaseManager } = require('../cache/databaseManager');
+const { databaseManager } = require('../database/databaseManager');
 const { preProcessMessage, isCommand, getExpiration } = require('../utils/baileys/messageHelper');
 const { sendOmniZapMessage, sendTextMessage, sendStickerMessage, sendReaction, formatErrorMessage } = require('../utils/messageUtils');
 const { COMMAND_PREFIX } = require('../utils/constants');
